src/pyExpandObjects/requirements.txt

# always needed
jsonschema==3.2.0
pyYaml==5.4.1

# for building sphinx
sphinx==3.4.3

# for local testing
flake8==3.8.4
nose==1.3.7
coverage==5.4

# for testing on CI so it can push to coveralls
coveralls==3.0.0
pandas==1.2.3
nose-exclude==0.5.0

# for building packages
pyinstaller==4.2
